backend/requirements.txt
# requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0